fix(api): Move constant 'ApiUrl' to transport.ts to avoid repetition
This commit moves the 'ApiUrl' constant from the API index to the Transport class to avoid repetition in the code and make it easier to manage the base URL for HTTP requests. It also renames 'ApiUrl' to 'ApiBaseUrl' in both files to give it a more accurate name in relation to its usage.