feat: add media item feature filter
This commit introduces a new feature filter which allows users to filter media items by feature, in particular, by favorites. A new class FeatureFilter is added along with its corresponding property and setters in Filters. This also exposes the new FeatureFilter class in `index.ts` which makes it available to use in other files. Finally, some modifications are made in `pack.ts` to use this new filter.