fix: make mediaItems and coverPhotoMediaItem in Album nullable
This commit improves the albumParser function to add a coverPhotoMediaItem field to Album object. Additionally, the mediaItems field in the Album object is made nullable since not all albums have media items. The related code changes include adding a coverPhotoMediaItem field in the return object, updating the coverPhotoMediaItem field after parsing, and changing the type of mediaItems. This change will improve the handling of null media items for albums.